+++ griptape/drivers/__init__.py
@@ -100,6 +100,7 @@ from .web_search.base_web_search_driver import BaseWebSearchDriver
 from .web_search.google_web_search_driver import GoogleWebSearchDriver
 from .web_search.duck_duck_go_web_search_driver import DuckDuckGoWebSearchDriver
 from .web_search.exa_web_search_driver import ExaWebSearchDriver
+from .web_search.tavily_web_search_driver import TavilyWebSearchDriver
 
 from .event_listener.base_event_listener_driver import BaseEventListenerDriver
 from .event_listener.amazon_sqs_event_listener_driver import AmazonSqsEventListenerDriver
@@ -215,6 +216,7 @@ __all__ = [
     "GoogleWebSearchDriver",
     "DuckDuckGoWebSearchDriver",
     "ExaWebSearchDriver",
+    "TavilyWebSearchDriver",
     "BaseEventListenerDriver",
     "AmazonSqsEventListenerDriver",
     "WebhookEventListenerDriver",