+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Added
 - Parameter `pipeline_task` on `HuggingFacePipelinePromptDriver` for creating different types of `Pipeline`s.
 
+- `TavilyWebSearchDriver` to integrate Tavily's web search capabilities.
 ### Changed
 - **BREAKING**: Renamed parameters on several classes to `client`:
   - `bedrock_client` on `AmazonBedrockCohereEmbeddingDriver`.
@@ -44,6 +45,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Ability to use Drivers Configs as Context Managers for temporarily setting the default Drivers.
 - Generic type support to `ListArtifact`.
 - Iteration support to `ListArtifact`.
+
 
 
 ### Changed